% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_bivarate.R
\name{freq_bivar_helper}
\alias{freq_bivar_helper}
\title{Frequence Helper Function
Helper Function, which takes various parameters as inputs. Parameters: Do weighting and outputs a reduced summary table (n)}
\usage{
freq_bivar_helper(df, var, byvar, summary, weights)
}
\arguments{
\item{df}{Input as Data.table}

\item{var}{Input as String}

\item{byvar}{Input as String}

\item{summary}{Summery: True/False}

\item{weights}{Input as String}
}
\value{
data.table
}
\description{
Frequence Helper Function
Helper Function, which takes various parameters as inputs. Parameters: Do weighting and outputs a reduced summary table (n)
}
\examples{
freq_bivar_helper(data, var = c("var1"), byvar = c("var2"), summary = T, weights = "gewicht")

}
