% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_bivarate.R
\name{tableband_bi}
\alias{tableband_bi}
\title{Create List of dataframes with bivarite frequency tables}
\usage{
tableband_bi(
  df,
  row_vars,
  col_vars,
  summary = F,
  var_labels = T,
  weights = NULL,
  col_per = F
)
}
\arguments{
\item{df}{Input as Dataframe}

\item{row_vars}{Vector of variable names for breaks}

\item{col_vars}{Vector of variable names}

\item{summary}{Output as summary (boolean variable)}

\item{var_labels}{Use of variable labels (set as attributes in Inputdataframe)}

\item{weights}{Variable name of weighting variable}

\item{col_per}{If False (Default): row percent, if True: column percent}
}
\value{
List of dataframes. Each col_var is a list element.
}
\description{
Create List of dataframes with bivarite frequency tables
}
\examples{
tableband_bi(pp2, row_vars = c("q1GEND", "q10KNOW_1"), col_vars = c("q10KNOW_5", "q10KNOW_4"), summary = T)

}
