% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_bivarate.R
\name{tableband_mean}
\alias{tableband_mean}
\title{Create a bivarate mean table with breaks}
\usage{
tableband_mean(
  df,
  row_vars,
  col_vars,
  weight = NULL,
  var_labels = F,
  digits = 2,
  summary = F,
  item_labels = F
)
}
\arguments{
\item{df}{Input: data.frame or data.table}

\item{row_vars}{Vector of variable names for breaks}

\item{col_vars}{Vector of variable names}

\item{weight}{Variable name of weighting variable}

\item{var_labels}{Use of variable labels (set as attributes in Inputdataframe)}

\item{digits}{Digits of mean}

\item{summary}{Output as summary (boolean)}

\item{item_labels}{Boolean parameter for showing item labels (Default: True) or variable names (False). The item labels are generated from attributes ("label").}
}
\description{
Create a bivarate mean table with breaks
}
\examples{
var_old <- paste0("q10KNOW_", c(1:3))
test <- tableband_mean(pp2, row_vars = c("q1GEND", "attribute_4"), col_var = c(var_old), summary = T, item_labels = T, var_labels = T)
}
