% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_univariate.R
\name{tableband_uni}
\alias{tableband_uni}
\title{Create a Univariate Frequency Table}
\usage{
tableband_uni(df, vars, var_labels = T, weights = NULL)
}
\arguments{
\item{df}{Dataframes}

\item{vars}{Vector of column names}

\item{weights}{Column name of weight vector}
}
\value{
Returns a dataframe with frequency counts (n, percent, valid percent)
}
\description{
Create a Univariate Frequency Table
}
