% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_univariate.R
\name{freq_univariate_helper}
\alias{freq_univariate_helper}
\title{Helper Function for Univariate Tables
Create a simple frequency table.}
\usage{
freq_univariate_helper(df, var)
}
\arguments{
\item{var}{}
}
\value{
data.table
}
\description{
Helper Function for Univariate Tables
Create a simple frequency table.
}
