% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_univariate.R
\name{tableband_uni}
\alias{tableband_uni}
\title{Create a Univariate Frequency Table}
\usage{
tableband_uni(df, vars)
}
\arguments{
\item{weights}{}
}
\description{
Create a Univariate Frequency Table
}
